requirements.txt
asgiref==3.5.2
Django==4.0.4
gunicorn==20.1.0
psycopg2==2.8.6
sqlparse==0.4.2
whitenoise==6.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.33.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.5.0
pytz==2023.3
requests==2.30.0
six==1.16.0
urllib3==1.26.15
